# schemas/openapi/parties.yaml
openapi: 3.1.0
info:
  title: CREST parties (né registry, #50)
  version: "1"
  description: |
    ## Who may call what (#102)

    Every service verifies bearer tokens; this section is who each surface
    answers. Four kinds of path:

    - **Party-authorized** — acts in a named party's name; the caller must be
      that party or hold act-for-party where the action is scoped
      (X-CREST-On-Behalf-Of names the party). Marked `security: bearerAuth`.
    - **Authenticated** — any signed-in caller (custodian queues, operations
      lists). Also `bearerAuth`.
    - **Open by design** — verification, status list, instance, capability
      reads by unguessable id, and the bootstrap plane; each carries a comment
      saying why.
    - **`/internal/*`** — the other services' doors, unauthenticated by the
      service-identity ruling (§16, 2026-08-25): service traffic lives on
      internal routes and fencing them to the service network is the
      deployment's job.
    Who a party is, what they agreed to, what they are authorised to do, and — the
    part the rest of the system leans on — whether an identifier from a source
    system resolves to exactly one of them.

    Request and response bodies that are CREST primitives are referenced by the
    `$id` of their JSON Schema under `schemas/`, never restated here. A second copy
    of a primitive is the thing `schemas/README.md` exists to prevent: two
    hand-maintained definitions agree right up until they don't.
servers:
...
paths:
...
  /v1/parties/{id}:
    get:
      operationId: getParty
      summary: Read a party
      description: Answers who this identifier belongs to, including their contact routes.
      parameters: [{ $ref: "#/components/parameters/PartyID" }]
      responses:
        "200":
          description: The party.
          content:
            application/json:
              schema:
                $ref: "urn:crest:schema:primitives:party:1"
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/Internal" }
...
  /v1/authorizations/permits:
    get:
      operationId: permits
      summary: May this party do this, here, at this moment?
      description: >
        A yes/no against the authorizations on file, evaluated at a point in time.
        `at` is a parameter rather than always-now so a caller can ask whether an
        action was permitted when it happened — an authorization that has since
        expired did not retroactively un-permit last month's work.
      parameters:
        - { name: partyId, in: query, schema: { type: string } }
        - { name: function, in: query, schema: { type: string } }
        - { name: contextId, in: query, schema: { type: string } }
        - name: at
          in: query
          schema: { type: string, format: date-time }
          description: RFC3339. Defaults to the service clock's now.
      responses:
        "200":
          description: The answer, and the instant it was evaluated at.
          content:
            application/json:
              schema:
                type: object
                required: [permitted, at]
                properties:
                  permitted: { type: boolean }
                  overdue:
                    type: boolean
                    description: >
                      Every authorization behind this yes is past its review-by
                      date (§16). Never changes `permitted` — "flag overdue, keep
                      working" — and a caller that ignores it loses nothing a
                      worker needed.
                  at: { type: string, format: date-time }
        "400":
          description: "`at` is not RFC3339."
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "500": { $ref: "#/components/responses/Internal" }
...
  /v1/organisations/{id}/registration:
    get:
      operationId: getRegistration
      summary: Where an organisation's application stands
      description: >
        The applicant-facing read: state, exact terms version accepted, decider
        and reason — plus the organisation's self-declared `attributes` (kind,
        sector, country, contact person; #168). Attributes are bounded strings
        from the Party's generic attributes map; the vocabulary is deployment
        configuration, and they are never published to the public log.
      parameters: [{ $ref: "#/components/parameters/PartyID" }]
      responses:
        "200":
          description: The registration, with the organisation's attributes.
          content:
            application/json:
              schema:
                type: object
                properties:
                  partyId: { type: string }
                  state: { type: string }
                  termsId: { type: string }
                  termsVersion: { type: integer }
                  acceptedAt: { type: string, format: date-time }
                  decidedBy: { type: string }
                  reason: { type: string }
                  appliedAt: { type: string, format: date-time }
                  attributes:
                    type: object
                    additionalProperties: { type: string }
        "404":
          description: No such organisation application.
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "500": { $ref: "#/components/responses/Internal" }
...
  /v1/projects/{id}/invitations:
    post:
      operationId: sendInvitation
      summary: Invite an organisation onto this project
      description: >
        Screens `g2_9`/`g2_10`. The invitation is the OFFER whose acceptance
        creates the partner grant `POST .../partner-grants` writes — same
        authority, same scope, same terms-as-ceiling rule, one transaction at
        acceptance. It is not a new primitive: the standing fact is the
        Authorization, and the invitation is the recorded offer-and-answer.


        The invitee must be an organisation Party, but its registration need
        not be decided yet — `g2_5`'s point is that registration stands alone,
        so an offer may arrive before or after the registry's decision.
        Acceptance, not sending, is where approval is checked. The offer must
        carry an end date, because the grant it becomes must.
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [partyId, functions, period]
              properties:
                partyId: { type: string }
                functions:
                  type: array
                  items: { type: string }
                  description: Opaque L2 strings; the core never reads the words.
                period:
                  type: object
                  description: Must carry `end` — an offer of a grant that ends.
                note: { type: string, maxLength: 2000 }
      responses:
        "201":
          description: The offer, SENT.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Invitation" }
        "403": { description: Not permitted to configure this project. }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { description: The named configurator has not accepted the handover. }
        "422": { description: Not an organisation, no functions, or no end date. }
    get:
      operationId: listProjectInvitations
      summary: The offers this project has made
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: state, in: query, schema: { type: string, enum: [SENT, ACCEPTED, DECLINED] } }
      responses:
        "200":
          description: This project's invitations.
          content:
            application/json:
              schema:
                type: object
                properties:
                  invitations: { type: array, items: { $ref: "#/components/schemas/Invitation" } }
                  count: { type: integer }
        "403": { description: Not permitted. }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/organisations/{id}/invitations:
    get:
      operationId: listOrganisationInvitations
      summary: The organisation's invitation inbox
      description: >
        `g2_9`'s source. Narrowed to the organisation and authorized against
        it, for the same reason `GET /v1/projects` narrows — who is courting
        whom is not a public listing.
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: "#/components/parameters/PartyID" }
        - { name: state, in: query, schema: { type: string, enum: [SENT, ACCEPTED, DECLINED] } }
      responses:
        "200":
          description: The invitations addressed to this organisation.
          content:
            application/json:
              schema:
                type: object
                properties:
                  invitations: { type: array, items: { $ref: "#/components/schemas/Invitation" } }
                  count: { type: integer }
        "403": { description: Not this organisation, and not permitted to act for it. }
  /v1/invitations/{id}:
    get:
      operationId: getInvitation
      summary: One offer, with its whole trail
      description: >
        The offer plus every event there has been — the questions included,
        because a question asked on the record is on the record. Readable by
        the invited organisation (which may hold nothing on the project yet)
        and by anybody the project's own view gate admits.
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200":
          description: The invitation and its events.
          content:
            application/json:
              schema:
                type: object
                properties:
                  invitation: { $ref: "#/components/schemas/Invitation" }
                  events: { type: array, items: { $ref: "#/components/schemas/InvitationEvent" } }
        "403": { description: Neither side of the offer. }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/invitations/{id}/decision:
    post:
      operationId: decideInvitation
      summary: Accept or decline an invitation
      description: >
        `g2_9`'s Accept and Decline. Declining records who and why and destroys
        nothing — it contests the offer, not the organisation's standing.


        Accepting is where the ordering gate lives (`g2_5`): the organisation
        must by now be APPROVED and hold accepted terms, and the offered
        functions must fit inside them — because acceptance creates the partner
        grant in the same transaction, riding the organisation's own accepted
        terms version, so there is no state in which the org said yes and holds
        nothing (`g2_10`). An undecided registration answers 409 with "not
        yet", never "no": the offer does not expire with the wait.
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision]
              properties:
                decision: { type: string, enum: [accepted, declined] }
                reason: { type: string, description: Required when declining. }
      responses:
        "200":
          description: The answered invitation; `grantId` names the grant an acceptance created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Invitation" }
        "400": { $ref: "#/components/responses/BadBody" }
        "403": { description: Not the invited organisation, and not permitted to act for it. }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: >
            Already answered, the registration is not yet APPROVED, or the
            organisation has accepted no terms.
        "422":
          description: >
            A decline with no reason, or an offer that outgrew the terms —
            every uncovered function named.
  /v1/invitations/{id}/questions:
    post:
      operationId: askInvitationQuestion
      summary: Ask a question on an open offer
      description: >
        `g2_9`'s third button. Either side — the invited organisation, or the
        project's owner or named configurator — may ask while the offer is
        open. A question changes no state; it is conversation on the record,
        appended to the trail. After the answer there is nobody holding the
        offer, so the trail closes with it.
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [askedBy, text]
              properties:
                askedBy: { type: string }
                text: { type: string, maxLength: 2000 }
      responses:
        "201":
          description: The recorded question.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InvitationEvent" }
        "400": { $ref: "#/components/responses/BadBody" }
        "403": { description: The asker is neither side of this offer. }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { description: The offer has been answered. }
  /v1/organisations/{id}/terms-requests:
    post:
      operationId: createTermsRequest
      summary: Open a request for wider terms
      description: >
        Screens `g2_6`/`g2_7`/`g2_11`. An APPROVED organisation asks to move to
        a wider *published* terms version — APPROVED only, deliberately:
        registration-time terms go through `.../terms-acceptance` and the
        registration decision, and this door is the one thing that path cannot
        do, which is move an already-approved organisation with review in
        between.


        Qualification documents are DECLARED, never uploaded: `{kind, ref,
        hash}` — the kind from the deployment's own taxonomy (L2, opaque
        here), the ref into its own document custody, bounded and single-line
        so nothing document-shaped can be smuggled in. This service has no
        blob store and grows none.
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/PartyID" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [termsId, termsVersion]
              properties:
                termsId: { type: string }
                termsVersion: { type: integer, minimum: 1 }
                documents:
                  type: array
                  items: { $ref: "#/components/schemas/DeclaredDocument" }
      responses:
        "201":
          description: The request, DRAFT.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TermsRequest" }
        "403": { description: Not this organisation, or its registration is not APPROVED. }
        "422": { description: No such published terms version, or an invalid document declaration. }
    get:
      operationId: listOrganisationTermsRequests
      summary: This organisation's terms requests
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: "#/components/parameters/PartyID" }
        - { name: state, in: query, schema: { type: string, enum: [DRAFT, SUBMITTED, WITHDRAWN, APPROVED, DENIED] } }
      responses:
        "200":
          description: The requests, newest state included.
          content:
            application/json:
              schema:
                type: object
                properties:
                  requests: { type: array, items: { $ref: "#/components/schemas/TermsRequest" } }
                  count: { type: integer }
        "403": { description: Not this organisation, and not permitted to act for it. }
  /v1/terms-requests:
    get:
      operationId: listTermsRequestQueue
      summary: The review queue
      description: >
        The queue `g2_8`'s "sent for review" lands in. Authenticated like the
        partner directory: who reviews is L2 routing this service does not
        encode. `state` defaults to SUBMITTED because the queue is the
        question this read answers.
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: state, in: query, schema: { type: string, enum: [DRAFT, SUBMITTED, WITHDRAWN, APPROVED, DENIED], default: SUBMITTED } }
      responses:
        "200":
          description: The matching requests.
          content:
            application/json:
              schema:
                type: object
                properties:
                  requests: { type: array, items: { $ref: "#/components/schemas/TermsRequest" } }
                  count: { type: integer }
        "401": { description: No verified caller. }
  /v1/terms-requests/{id}:
    get:
      operationId: getTermsRequest
      summary: One request, whole
      description: >
        State, declared documents, the append-only trail and every check
        verdict — `g2_8`'s status view and `g2_12`'s "what is checked" on one
        read.
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200":
          description: The request, its events and its checks.
          content:
            application/json:
              schema:
                type: object
                properties:
                  request: { $ref: "#/components/schemas/TermsRequest" }
                  events: { type: array, items: { $ref: "#/components/schemas/TermsRequestEvent" } }
                  checks: { type: array, items: { $ref: "#/components/schemas/CheckVerdict" } }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/terms-requests/{id}/documents:
    put:
      operationId: replaceRequestDocuments
      summary: Save the draft's declared documents
      description: >
        `g2_7`'s "Save draft". DRAFT only — after submission the reviewer must
        be looking at what was submitted.
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                documents:
                  type: array
                  items: { $ref: "#/components/schemas/DeclaredDocument" }
      responses:
        "200":
          description: The draft, with its new list.
          content: { application/json: { schema: { $ref: "#/components/schemas/TermsRequest" } } }
        "403": { description: Not this organisation. }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { description: Not a draft any more. }
        "422": { description: An invalid document declaration. }
  /v1/terms-requests/{id}/submit:
    post:
      operationId: submitTermsRequest
      summary: Send the request for review
      description: >
        `g2_7`'s "Submit" → `g2_8`. The declared documents freeze and the
        request enters the review queue. Whether they are ENOUGH is the
        review's question, under L2 policy; whether the request is a
        submittable thing is this endpoint's.
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200":
          description: SUBMITTED, with the submitter and time on it.
          content: { application/json: { schema: { $ref: "#/components/schemas/TermsRequest" } } }
        "403": { description: Not this organisation. }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { description: Already submitted, or already settled. }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        An OIDC access token from the deployment's identity provider (#89, §4.1).
        In a deployment that is eSignet; locally it is tools/mocks/oidc, which
        mints real ES256 tokens against a real JWKS so the verification path
        under test is the production one.


        The provider's `sub` is never used directly. CREST re-salts it under a
        per-deployment key before anything stores or compares it, because #52
        and #63 are open against eSignet's pairwise subject — it partitions by
        relying party rather than by client, and derives without a
        per-deployment salt. Two deployments reading the same eSignet therefore
        cannot correlate their workers by subject.
  parameters:
...
    PartyID:
      name: id
      in: path
      required: true
      schema: { type: string }
      description: The party identifier.
...
  responses:
...
    NotFound:
      description: No such record.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...